Add files via upload
The top row of small black button was my first try. It is non viable. The numpad is a much better solution. The numpat is connested through a resistor ladder. The 4 row LCD screen has no complete Swedish character set, and the character that exist is odd looking. The microcontoller by itself can tolerate rather noisy environment, but the screen can not, so the screen needs to bee rewritten often,have its own power (not going through the microcontroller-board) and some extra capacitors.